Crate narwhal_worker

Modules

metrics

Structs

LazyNarwhalClient

TODO: add NarwhalClient trait and implement RemoteNarwhalClient with grpc. A Narwhal client that instantiates LocalNarwhalClient lazily.

LocalNarwhalClient

A client that connects to Narwhal locally.

TrivialTransactionValidator

Simple validator that accepts all transactions and batches.

Worker

Constants

NUM_SHUTDOWN_RECEIVERS

The number of shutdown receivers to create on startup. We need one per component loop.

Traits

TransactionValidator

Defines the validation procedure for receiving either a new single transaction (from a client) of a batch of transactions (from another validator). Invalid transactions will not receive further processing.