Trait narwhal_worker::TransactionValidator

pub trait TransactionValidator: Clone + Send + Sync + 'static {
    type Error: Display + Debug + Send + Sync + 'static;

    // Required methods
    fn validate(&self, t: &[u8]) -> Result<(), Self::Error>;
    fn validate_batch(
        &self,
        b: &Batch,
        protocol_config: &ProtocolConfig
    ) -> Result<(), Self::Error>;
}

Defines the validation procedure for receiving either a new single transaction (from a client) of a batch of transactions (from another validator). Invalid transactions will not receive further processing.

Required Associated Types

type Error: Display + Debug + Send + Sync + 'static

Required Methods

fn validate(&self, t: &[u8]) -> Result<(), Self::Error>

Determines if a transaction valid for the worker to consider putting in a batch

fn validate_batch( &self, b: &Batch, protocol_config: &ProtocolConfig ) -> Result<(), Self::Error>

Determines if this batch can be voted on

Object Safety

This trait is not object safe.